split fdp into device
